Move common pagination params to shared method
